refactor(openapi): buildV3.ts
